member

+++ src/main/java/com/example/shoppingweb/repository/CategoryRepository.java
@@ -3,5 +3,5 @@ package com.example.shoppingweb.repository;
 import com.example.shoppingweb.model.Cart;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepository extends JpaRepository<Cart, Long> {
+public interface CategoryRepository extends JpaRepository<Cart, Integer> {
 }
